+++ bridge/opencensus/internal/oc2otel/span_context.go
@@ -14,9 +14,19 @@ func SpanContext(sc octrace.SpanContext) trace.SpanContext {
 	if sc.IsSampled() {
 		traceFlags = trace.FlagsSampled
 	}
+
+	entries := slices.Clone(sc.Tracestate.Entries())
+	slices.Reverse(entries)
+
+	tsOtel := trace.TraceState{}
+	for _, entry := range entries {
+		tsOtel, _ = tsOtel.Insert(entry.Key, entry.Value)
+	}
+
 	return trace.NewSpanContext(trace.SpanContextConfig{
 		TraceID:    trace.TraceID(sc.TraceID),
 		SpanID:     trace.SpanID(sc.SpanID),
 		TraceFlags: traceFlags,
+		TraceState: tsOtel,
 	})
 }